Removing {failed,completed,new} proposal directories
Cleaning up the proposal directory. In lieu of the current layout, we'll merge
completed & accepted proposals and store them in the root of /proposals, while
non-accepted proposals will not be merged. Historical metadata (e.g., reasons
for rejected proposal, discussion during the vetting phase, etc.) can be found
on GitHub.

We can revisit organization as needed in the future.

Signed-off-by: clouderati <35942204+clouderati@users.noreply.github.com>