Removing myself from MAINTAINERS.md
I am stepping down as a maintainer of Harbor. I haven't been active in
the past few months, I'm also changing employers in about a week and I
do not anticipate being able to be more active in the near future.

It's been a great experience working with you all and I hope to
contribute again in the future.  I wish you all the best of luck in the
years to come.

Signed-off-by: Nathan Lowe <public@nlowe.me>